Refactored: - Views and Widget Ids to follow the convention of: widgetNameBtn, viewNameBtn. - Moved Activities into sub packages. - Miscellaneous cleaning up and naming.
Additional Functionality:
- Added base emergency contact functionality: creation and viewing of contacts.
- Added current_user table, which will be used to track the current user.

Details about user system and current_user table:
- current_user table should only contain 1 row: the active user.
- This single row in the current_user table (user_id and date_time) is used to track the current user, to get a user's data from the contacts table etc by its user_id.
- An alternative option was to store the user_id in SharedPreferences, but this is a security/ privacy risk.

How it works:
When the user logs into the system, they are added as the *only* value in current_user table.
As long as this table is cleared just before logging the user in, after logout and after app destroy, this should keep the table to a single row- the currently active user within the application.
This allows us to query data from database, using the current user's user id, getting a user's emergency contacts etc.